pyproject.toml
[project]
name = "crm-sd-dmfa"
version = "0.4.3"
description = "Coupled SD (BPTK-Py) + dMFA (flodym) scaffold for CRM circularity/criticality/resilience"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "typer>=0.9",
]

[project.optional-dependencies]
model = [
  "bptk-py",
  "flodym",
]

[project.scripts]
crm-model = "crm_model.cli:app"
